ref(cmd): break up command definitions
