+++ cmd/helm/doctor.go
@@ -1,12 +1,26 @@
 package main
 
 import (
+	"github.com/codegangsta/cli"
 	"github.com/deis/helm-dm/dm"
 	"github.com/deis/helm-dm/format"
 	"github.com/deis/helm-dm/kubectl"
 )
 
-func doctor() error {
+func init() {
+	addCommands(doctorCmd())
+}
+
+func doctorCmd() cli.Command {
+	return cli.Command{
+		Name:      "doctor",
+		Usage:     "Run a series of checks for necessary prerequisites.",
+		ArgsUsage: "",
+		Action:    func(c *cli.Context) { run(c, doctor) },
+	}
+}
+
+func doctor(c *cli.Context) error {
 	var runner kubectl.Runner
 	runner = &kubectl.RealRunner{}
 	if dm.IsInstalled(runner) {
